requirements.txt
argparse==1.2.1
distribute==0.6.31
wsgiref==0.1.2
Django==1.4.3
django-grappelli==2.4.3
mimeparse>=0.1.3
python-dateutil>=2.1
PyYAML
python-digest
biplist
django-tastypie==0.9.11
psycopg2
dj-database-url